Second Auth step successfuly implemented
It has been added a couple of recurrent texts
to the file Constants this in order to have a
better control on the texts, the other changes
where because I found a couple of issues trying
to decrypt the sensitive data, but is solved
