+++ app/src/main/java/com/example/testtimetonic/MainActivity.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.testtimetonic.Model.Crythographer.clearFilesAndKeys
 import com.example.testtimetonic.ModelViews.LoginVM
 import com.example.testtimetonic.ui.login.LoginView
 import com.example.testtimetonic.ui.theme.TestTimeTonicTheme
@@ -21,6 +22,8 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        clearFilesAndKeys(context = baseContext)
 
         setContent {
             val loginViewModel = LoginVM(application)
