+++ app/src/main/java/com/example/testtimetonic/Model/Constants.kt
@@ -1,0 +1,7 @@
+package com.example.testtimetonic.Model
+
+enum class Constants(val constanVal: String) {
+    CREATE_OAUTHKEY("createOauthkey"),
+    STATUS_OK("ok"),
+    STATUS_NOT_OK("nok")
+}
